+++ man-roxygen/references.R
@@ -1,0 +1,17 @@
+#' @references Anderson, S. F., & Kelley, K. (2024). Sample size planning for
+#'   replication studies: The devil is in the design. \emph{Psychological
+#'   Methods, 29,} 844--867. \doi{10.1037/met0000520}
+#'
+#'   Anderson, S. F., Kelley, K., & Maxwell, S. E. (2017). Sample-size planning
+#'   for more accurate statistical power: A method adjusting sample effect sizes
+#'   for publication bias and uncertainty. \emph{Psychological Science, 28,}
+#'   1547--1562. \doi{10.1177/0956797617723724}
+#'
+#'   Anderson, S. F., & Maxwell, S. E. (2017). Addressing the 'replication
+#'   crisis': Using original studies to design replication studies with
+#'   appropriate statistical power. \emph{Multivariate Behavioral Research, 52,}
+#'   305--324.
+#'
+#'   Taylor, D. J., & Muller, K. E. (1996). Bias in linear model power and
+#'   sample size calculation due to estimating noncentrality. \emph{Communications
+#'   in Statistics: Theory and Methods, 25,} 1595--1610.